+++ eslint.config.mjs
@@ -1,7 +1,7 @@
 // @ts-check
 import eslint from "@eslint/js";
 import tseslint from "typescript-eslint";
-import eslintConfigPrettier from "eslint-config-prettier";
+import eslintConfigPrettier from "eslint-plugin-prettier/recommended";
 
 export default tseslint.config(
 	eslint.configs.recommended,