+++ packages/vscode-css-languageservice/src/test/sass/sassNavigation.test.ts
@@ -902,6 +902,31 @@ foo
 			);
 		});
 
+		// Slightly different from aliases, loadPaths act as additional "roots" for import strings.
+		test("load paths", async () => {
+			const fixtureRoot = path.resolve(__dirname, "../../../../src/test/scss/linkFixture/loadPaths");
+			const getDocumentUri = (relativePath: string) => {
+				return URI.file(path.resolve(fixtureRoot, relativePath)).toString(true);
+			};
+
+			const settings = {
+				loadPaths: ["shared/"],
+			};
+
+			await assertDynamicLinks(
+				getDocumentUri("./"),
+				`@import 'my-lib/variables'`,
+				[
+					{
+						range: newRange(8, 26),
+						target: getDocumentUri("./shared/my-lib/variables.scss"),
+						type: nodes.NodeType.Import,
+					},
+				],
+				settings,
+			);
+		});
+
 		test("Sass aliased links", async function () {
 			const fixtureRoot = path.resolve(__dirname, "../../../src/test/sass/linkFixture");
 			const getDocumentUri = (relativePath: string) => {